Refactor map generator
